requirements.txt
redis
mysql-connector-python
psutil
pyinstaller
fastapi
uvicorn
typing
pydantic